handlers.tsを利用して循環参照を解決
