+++ src/game/handler/handlers.ts
@@ -1,0 +1,4 @@
+// A trick to avoid circular import
+export * from './CountryHandler';
+export * from './CountryAIHanlder';
+export * from './CountryPlayerHandler';
