+++ src/game/scene/MainScene.ts
@@ -3,10 +3,9 @@ import { MapViewport } from '../container/MapViewport';
 import type { Country } from '../data/Country';
 import type { Province } from '../data/Provice';
 import { data } from '../GameManager';
+import { CountryPlayerHandler, eventHandler } from '../handler/handlers';
 import { Observable } from '../util/Observable';
 import { Scene } from './Scene';
-import type { eventHandler } from '../handler/CountryHandler';
-import { CountryPlayerHandler } from '../handler/CountryHandler';
 
 export class MainScene extends Scene {
   private readonly map: MapViewport;
